backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.10
python-dotenv==1.0.1
pydantic[email]==2.8.2
PyJWT==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9